+++ lib/src/will_mixins/will_cancel/will_cancel_mixin.dart
@@ -12,10 +12,8 @@
 
 import 'dart:async' show FutureOr;
 
-import 'package:flutter/foundation.dart'
-    show kDebugMode, mustCallSuper, nonVirtual;
-
-import '/src/_utils/_index.g.dart';
+import 'package:df_type/df_type.dart' show FutureOrController;
+import 'package:flutter/foundation.dart' show kDebugMode, mustCallSuper, nonVirtual;
 
 // ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
@@ -28,8 +26,7 @@ import '/src/_utils/_index.g.dart';
 /// invoked on each resource wrapped with [willCancel].
 mixin WillCancelMixin on CancelMixin {
   /// The list of resources marked for cancel via [willCancel].
-  Set<_ToCancelResource<dynamic>> get toCancelResources =>
-      Set.unmodifiable(_toCancelResources);
+  Set<_ToCancelResource<dynamic>> get toCancelResources => Set.unmodifiable(_toCancelResources);
 
   final Set<_ToCancelResource<dynamic>> _toCancelResources = {};
 
@@ -46,18 +43,16 @@ mixin WillCancelMixin on CancelMixin {
   ///
   /// Returns the resource back to allow for easy chaining or assignment.
   @nonVirtual
-  T willCancel<T>(T resource, {OnBeforeCallback<T>? onBeforeCancel}) {
+  T willCancel<T>(T resource, {_OnBeforeCallback<T>? onBeforeCancel}) {
     // Verify that the resource has a cancel method in debug mode.
     _verifyCancelMethod(resource);
     final disposable = (
       resource: resource as dynamic,
-      onBeforeCancel:
-          onBeforeCancel != null ? (dynamic e) => onBeforeCancel(e as T) : null,
+      onBeforeCancel: onBeforeCancel != null ? (dynamic e) => onBeforeCancel(e as T) : null,
     );
 
     // Check for any duplicate resource.
-    final duplicate =
-        _toCancelResources.where((e) => e.resource == resource).firstOrNull;
+    final duplicate = _toCancelResources.where((e) => e.resource == resource).firstOrNull;
 
     if (duplicate != null) {
       if (kDebugMode) {
@@ -79,11 +74,11 @@ mixin WillCancelMixin on CancelMixin {
   @mustCallSuper
   @override
   FutureOr<void> cancel() {
-    final fom = FutureOrManager();
+    final foc = FutureOrController();
 
     try {
       // Call the parent's cancel method.
-      fom.add(super.cancel());
+      foc.add(super.cancel());
 
       for (final disposable in _toCancelResources) {
         final resource = disposable.resource;
@@ -93,13 +88,13 @@ mixin WillCancelMixin on CancelMixin {
         // Attempt to call onBeforeCancel, catching and copying any exceptions.
         Object? onBeforeCancelError;
         try {
-          fom.add(disposable.onBeforeCancel?.call(resource));
+          foc.add(disposable.onBeforeCancel?.call(resource));
         } catch (e) {
           onBeforeCancelError = e;
         }
 
         // Attempt to call cancel on the resource.
-        fom.add(resource.cancel());
+        foc.add(resource.cancel());
 
         // If successful, rethrow any exception from onBeforeCancel.
         if (onBeforeCancelError != null) {
@@ -109,11 +104,11 @@ mixin WillCancelMixin on CancelMixin {
     } catch (e) {
       // Collect exceptions to throw them all at the end, ensuring cancel gets
       // called on all resources.
-      fom.addException(e);
+      foc.addException(e);
     }
 
     // Return a Future or complete synchronously.
-    return fom.complete();
+    return foc.complete();
   }
 
   /// Throws [NoCancelMethodDebugError] if [resource] does not have a `cancel`
@@ -129,20 +124,13 @@ mixin WillCancelMixin on CancelMixin {
   static bool hasValidCancelMethod(dynamic resource) {
     try {
       final method = resource.cancel;
-      final isValid = method is FutureOrCallback;
+      final isValid = method is _FutureOrCallback;
       return isValid;
     } on NoSuchMethodError {
       return false;
     }
   }
 }
-
-// ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
-
-typedef _ToCancelResource<T> = ({
-  T resource,
-  OnBeforeCallback<T>? onBeforeCancel,
-});
 
 // ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
@@ -160,7 +148,7 @@ final class NoCancelMethodDebugError extends Error {
   String toString() {
     return '[$NoCancelMethodDebugError] The type $resourceType cannot be used '
         'with willCancel() as it has no "cancel" method or one that conforms '
-        'to $FutureOrCallback.';
+        'to $_FutureOrCallback.';
   }
 }
 
@@ -186,3 +174,13 @@ mixin CancelMixin {
   /// Override to define the cancel operation.
   FutureOr<void> cancel();
 }
+
+// ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+
+typedef _ToCancelResource<T> = ({
+  T resource,
+  _OnBeforeCallback<T>? onBeforeCancel,
+});
+
+typedef _FutureOrCallback<T> = FutureOr<void> Function();
+typedef _OnBeforeCallback<T> = FutureOr<void> Function(T resource);